requirements.txt
Flask==3.1.3
groq==1.1.1
python-dotenv==1.2.2
Flask-Login==0.6.3
Werkzeug==3.1.6
